Refactor Data Layer
